Integrate with job-controller-service
Triggering is now done on a task instead of a submission status or the
old task system. New tasks are also generated and their status is
updated according to the job-controller model.

Some functions about deletions of submissions have been removed as they
could not have been executed (no config for this in the delta-notifier).
Insertions are now also less filtered. Not sure if this is still exactly
what should happen. (E.g. we don't filter on the status of the
submission, because the old triggers where SENT_STATUS on the submission
or CONCEPT_STATUS on the old task, and these statuses are the only
statuses that will result in a successful validation.)